cloudFunctions/requirements.txt
# Function dependencies, for example:
# package>=version
websocket_client >=0.53.0
simplejson >=3.17.0